src/main/java/com/github/brunomndantas/jscrapper/support/parser/collection/text/reference/CollectionReferenceIntegerTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.collection.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.collection.text.CollectionTextParser;
import org.openqa.selenium.WebElement;

public class CollectionReferenceIntegerTextParser extends CollectionTextParser {

    public CollectionReferenceIntegerTextParser() {
        super(null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Integer.parseInt(element.getText());
    }

}
