src/test/java/com/github/brunomndantas/jscrapper/support/parser/collection/text/reference/CollectionReferenceIntegerTextParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.collection.text.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class CollectionReferenceIntegerTextParserTest {

    @Test
    public void constructorTest() {
         CollectionReferenceIntegerTextParser parser = new CollectionReferenceIntegerTextParser();
        assertEquals(null, parser.getDefaultValue());
    }

    @Test
    public void parseElement() throws Exception {
         CollectionReferenceIntegerTextParser parser = new CollectionReferenceIntegerTextParser();

        assertEquals(2, parser.parseElement(new DummyElement(){
            @Override
            public String getText() {
                return "2";
            }
        }));
    }

}
